refactor(rktio): move connector layer primitives to rktio itself
