+++ pycket/rktio/rktio.py
@@ -1,18 +1,257 @@
-import os
+from pycket import values
+from pycket import vector as values_vector
+from pycket import values, values_string
+from pycket.prims.expose import expose, expose_val
 
-from rpython.translator.tool.cbuild import ExternalCompilationInfo
+from rpython.rtyper.lltypesystem import rffi
 
 from pycket.rktio.types import *
 from pycket.rktio._rktio_bootstrap import *
+from pycket.rktio.bootstrap_structs import *
 
-# Load the librktio.a
-# TODO: make this absolute (pycket/rktio), instead of "this file"
-RKTIO_DIR = os.path.dirname(os.path.abspath(__file__))
-librktio_a = ExternalCompilationInfo(
-    includes=['rktio.h'],
-    include_dirs=[RKTIO_DIR],
-    libraries=['rktio'],
-    library_dirs=[RKTIO_DIR],
+
+###############################################
+############## connector layer ################
+###############################################
+"""
+This layer implements the following functions that are
+not coming from the librktio.a, and need to be manually
+added.
+
+They use the struct pointers defined in bootstrap_structs,
+and they use c_rffi primitives declared at _rktio_bootstrap.
+
+rktio_NULL
+rktio_filesize_ref
+rktio_timestamp_ref
+rktio_is_timestamp
+rktio_recv_length_ref
+rktio_recv_address_ref
+rktio_stat_to_vector
+rktio_identity_to_vector
+rktio_seconds_to_date*
+rktio_convert_result_to_vector
+rktio_to_bytes
+rktio_to_bytes_list
+rktio_to_shorts
+rktio_from_bytes_list
+rktio_free_bytes_list
+rktio_make_sha1_ctx
+rktio_make_sha2_ctx
+rktio_process_result_stdin_fd
+rktio_process_result_stdout_fd
+rktio_process_result_stderr_fd
+rktio_process_result_process
+rktio_status_running
+rktio_status_result
+rktio_pipe_results
+rktio_do_install_os_signal_handler
+rktio_get_ctl_c_handler
+
+"""
+
+def _wrap_int(val):
+    """Return the smallest Pycket integer that can hold *val*."""
+    try:
+        return values.W_Fixnum(val)
+    except OverflowError:
+        return values.W_Bignum.fromint(val)
+
+expose_val("rktio_NULL", values.w_false)
+
+@expose("rktio_filesize_ref", [W_RKTIO_FILESIZE_PTR])
+def rktio_filesize_ref(w_filesize_ptr):
+    r_filesize_ptr = rffi.cast(RKTIO_FILESIZE_PTR, w_filesize_ptr.to_rffi())
+    val = r_filesize_ptr[0]
+    if not val:
+        return values.w_false
+    return _wrap_int(val)
+
+@expose("rktio_timestamp_ref", [W_RKTIO_TIMESTAMP_PTR])
+def rktio_timestamp_ref(w_timestamp_ptr):
+    r_timestamp_ptr = rffi.cast(RKTIO_TIMESTAMP_PTR, w_timestamp_ptr.to_rffi())
+    val = r_timestamp_ptr[0]
+    return _wrap_int(val)
+
+
+@expose("rktio_is_timestamp", [values.W_Object])
+def rktio_is_timestamp(w_v):
+    """
+    Return #t iff `v` fits in the signed range of rktio_timestamp_t
+    """
+    if not isinstance(w_v, values.W_Integer):
+        return values.w_false
+
+    # Try to obtain a plain Python int without overflow.
+    try:
+        v = w_v.toint()
+    except OverflowError:
+        # Too large to fit in a machine int ⇒ definitely outside timestamp range
+        return values.w_false
+
+    bits_in_timestamp = rffi.sizeof(RKTIO_TIMESTAMP_T) * 8 # 32 or 64
+    radix = 1 << (bits_in_timestamp - 1) # 2^(n-1)
+
+    if -radix <= v < radix:
+        return values.w_true
+    else:
+        return values.w_false
+
+@expose("rkito_recv_length_ref", [W_RKTIO_LENGTH_AND_ADDRINFO_PTR])
+def rktio_recv_length_ref(w_len_and_addrinfo_ptr):
+    r_struct_ptr = rffi.cast(RKTIO_LENGTH_AND_ADDRINFO_PTR,
+                             w_len_and_addrinfo_ptr.to_rffi())
+    return _wrap_int(r_struct_ptr.c_len)
+
+@expose("rktio_recv_address_ref", [W_RKTIO_LENGTH_AND_ADDRINFO_PTR])
+def rktio_recv_address_ref(w_len_and_addrinfo_ptr):
+    ll_ptr = rffi.cast(RKTIO_LENGTH_AND_ADDRINFO_PTR,
+                       w_len_and_addrinfo_ptr.to_rffi())
+
+    charp = ll_ptr.c_address[0]
+
+    if not charp:
+        return values.w_false
+
+    # copy into a Pycket/Racket string object
+    py_str = rffi.charp2str(charp) # assumes utf-8
+    return values_string.W_String.fromstr_utf8(py_str)
+
+@expose("rktio_stat_to_vector", [W_RKTIO_STAT_PTR])
+def rktio_stat_to_vector(w_stat_ptr):
+    ll_ptr = rffi.cast(RKTIO_STAT_PTR, w_stat_ptr.to_rffi())
+
+    elems = [
+        _wrap_int(ll_ptr.c_device_id),
+        _wrap_int(ll_ptr.c_inode),
+        _wrap_int(ll_ptr.c_mode),
+        _wrap_int(ll_ptr.c_hardlink_count),
+        _wrap_int(ll_ptr.c_user_id),
+        _wrap_int(ll_ptr.c_group_id),
+        _wrap_int(ll_ptr.c_device_id_for_special_file),
+        _wrap_int(ll_ptr.c_size),
+        _wrap_int(ll_ptr.c_block_size),
+        _wrap_int(ll_ptr.c_block_count),
+
+        _wrap_int(ll_ptr.c_access_time_seconds),
+        _wrap_int(ll_ptr.c_access_time_nanoseconds),
+        _wrap_int(ll_ptr.c_modify_time_seconds),
+        _wrap_int(ll_ptr.c_modify_time_nanoseconds),
+        _wrap_int(ll_ptr.c_ctime_seconds),
+        _wrap_int(ll_ptr.c_ctime_nanoseconds),
+
+        values.w_true if ll_ptr.c_ctime_is_change_time else values.w_false,
+    ]
+    return values_vector.W_Vector.fromelements(elems)
+
+@expose("rktio_identity_to_vector", [W_RKTIO_IDENTITY_PTR])
+def rktio_identity_to_vector(w_id_ptr):
+    ll_ptr = rffi.cast(RKTIO_IDENTITY_PTR, w_id_ptr.to_rffi())
+
+    elems = [
+        _wrap_int(ll_ptr.c_a),          # uintptr_t
+        _wrap_int(ll_ptr.c_b),
+        _wrap_int(ll_ptr.c_c),
+        _wrap_int(ll_ptr.c_a_bits),     # plain int -> still wrap
+        _wrap_int(ll_ptr.c_b_bits),
+        _wrap_int(ll_ptr.c_c_bits),
+    ]
+
+    return values_vector.W_Vector.fromelements(elems)
+
+"""
+@expose(
+    "rktio_seconds_to_date*",
+    [W_RKTIO_T_PTR, values.W_Integer, values.W_Integer, values.W_Integer],
 )
+def rktio_seconds_to_date_star(w_rktio, w_seconds, w_nanosec, w_get_gmt):
+    rktio_ptr = rffi.cast(RKTIO_T_PTR, w_rktio.to_rffi())
+
+    seconds  = rffi.cast(RKTIO_TIMESTAMP_T, w_seconds.toint())
+    nanosec  = rffi.cast(rffi.INT,          w_nanosec.toint())
+    get_gmt  = rffi.cast(rffi.INT,          w_get_gmt.toint())
+
+    date_p = c_rktio_seconds_to_date(rktio_ptr, seconds, nanosec, get_gmt)
+
+    # return #(errno symbol) when rktio_seconds_to_date fails
+    if not date_p:
+        err = rffi.get_saved_errno()
+        return values_vector.W_Vector.fromelements([
+            _wrap_int(err),
+            errno_code_to_symbol(err),            # returns a W_Symbol
+        ])
+
+    # success, build date*
+    d = date_p[0]
+
+    zone_name_val = (
+        values.w_false if not d.c_zone_name else
+        values_string.W_String.fromstr_utf8(rffi.charp2str(d.c_zone_name))
+    )
+
+    from pycket.prims.general import date_star_struct
+    # get and call the contstructor with these
+    # self.struct_type().constructor.call_with_extra_info(args...)
+    # will need env and cont to call this.
+
+
+    result = make_date_star(
+        _wrap_int(d.c_second),
+        _wrap_int(d.c_minute),
+        _wrap_int(d.c_hour),
+        _wrap_int(d.c_day),
+        _wrap_int(d.c_month),
+        _wrap_int(d.c_year),
+        _wrap_int(d.c_nanosecond),
+        _wrap_int(d.c_zone_offset),
+        values.w_true  if d.c_is_dst else values.w_false,
+        _wrap_int(d.c_day_of_week),
+        _wrap_int(d.c_day_of_year),
+        zone_name_val,
+    )
+
+    # free allocation performed by C
+    c_rktio_free(rktio_ptr, rffi.cast(rffi.VOIDP, date_p))
+
+    return result
+"""
+
+###############################################
+############## process_result #################
+
+def _wrap_process_null(addr):
+    return values.w_false if not addr else W_RKTIO_FD_T_PTR(addr)
+
+@expose("rktio_process_result_stdin_fd",
+        [W_RKTIO_PROCESS_RESULT_PTR])
+def rktio_process_result_stdin_fd(res_ptr):
+    res_ll = rffi.cast(RKTIO_PROCESS_RESULT_PTR, res_ptr.to_rffi())
+
+    greet()  # prints: I am function greet!
+
+    return _wrap_process_null(res_ll.c_stdin_fd)
+
+@expose("rktio_process_result_stdout_fd",
+        [W_RKTIO_PROCESS_RESULT_PTR])
+def rktio_process_result_stdout_fd(res_ptr):
+    res_ll = rffi.cast(RKTIO_PROCESS_RESULT_PTR, res_ptr.to_rffi())
+    return _wrap_process_null(res_ll.c_stdout_fd)
+
+@expose("rktio_process_result_stderr_fd",
+        [W_RKTIO_PROCESS_RESULT_PTR])
+def rktio_process_result_stderr_fd(res_ptr):
+    res_ll = rffi.cast(RKTIO_PROCESS_RESULT_PTR, res_ptr.to_rffi())
+    return _wrap_process_null(res_ll.c_stderr_fd)
+
+@expose("rktio_process_result_process",
+        [W_RKTIO_PROCESS_RESULT_PTR])
+def rktio_process_result_process(res_ptr):
+    res_ll = rffi.cast(RKTIO_PROCESS_RESULT_PTR, res_ptr.to_rffi())
+    return W_RKTIO_PROCESS_T_PTR(res_ll.c_process)   # never NULL, so no _wrap
+
+
+###############################################
+###############################################
+###############################################
